feature: create cars-availability-spec
